Splits useCart from CartContext for Fast Refresh functionality
