+++ src/components/CategorySelector/index.styled.jsx
@@ -46,4 +46,3 @@ export const TagSelect = styled.select`
     border-radius: 5px;
   }
 `;
-
