+++ src/components/OnlineShop/index.styled.jsx
@@ -2,6 +2,7 @@ import styled from "styled-components";
 
 export const OnlineShopContainer = styled.div`
   margin: 0 auto;
+  width: 100%;
 `;
 
 export const LoaderContainer = styled.div`
