+++ src/components/SortAndFilterContainer/index.styled.jsx
@@ -1,9 +1,20 @@
 import styled from "styled-components";
 
+export const SortAndFilterWrapperStyled = styled.div`
+  /* display: flex;
+  flex-wrap: nowrap;
+  width: 100%;
+  align-items: center;
+  margin: 0 auto;
+  color: ${(props) => props.theme.h2};
+  background-color: ${(props) => props.theme.header}; */
+`;
+
 export const SortAndFilterContainerStyled = styled.div`
   display: flex;
   flex-wrap: wrap;
   flex-direction: column;
+  width: 100%;
   justify-content: space-evenly;
   color: ${(props) => props.theme.h2};
   background-color: ${(props) => props.theme.header};
