+++ src/contexts/CartContext/index.jsx
@@ -1,11 +1,13 @@
 import PropTypes from "prop-types"; // Corrected import for PropTypes
-import { createContext, useContext, useState, useEffect } from "react";
-import { load, save } from "../../components/StorageUtils"; // Adjusted the import path
+import { createContext, useState, useEffect } from "react";
+import { load, save } from "../../components/StorageUtils";
 
-const CartContext = createContext();
 
-// eslint-disable-next-line react-refresh/only-export-components
-export const useCart = () => useContext(CartContext);
+
+export const CartContext = createContext();
+
+
+// export const useCart = () => useContext(CartContext);
 
 export const CartProvider = ({ children }) => {
   const [cartItems, setCartItems] = useState(() => load("cartItems") || []);
