+++ src/contexts/CartContext/useCart.js
@@ -1,0 +1,4 @@
+import { useContext } from "react";
+import { CartContext } from "../CartContext";
+
+export const useCart = () => useContext(CartContext);
